pyproject.toml
[build-system]
requires = [
    "hatchling",
    "hatch-vcs",
]
build-backend = "hatchling.build"

[project]
name = "pixelmatch-fast"
dynamic = ["version"]
description = "High-performance Python port of mapbox/pixelmatch for perceptual image comparison"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [{name = "Justus Rijke", email="justusrijke@gmail.com"}]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "click>=8",
    "numba>=0.57",
    "pillow>=10",
]

[project.urls]
Homepage = "https://github.com/JustusRijke/pixelmatch-fast"
Issues = "https://github.com/JustusRijke/pixelmatch-fast/issues"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.scripts]
pixelmatch = "pixelmatch.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/pixelmatch"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
local_scheme = "no-local-version"

[tool.pytest.ini_options]
addopts = ["--strict-config"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
extend-select = [
  "I",           # isort
  "NPY",         # NumPy specific rules
]
